refactor: :recycle: remove data folder
